# backend/counters-api/openapi.yaml
openapi: 3.1.0
info:
  title: Counters API
  version: v1
servers:
- url: /
paths:
  /api/counters:
    get:
      tags:
      - Counters
      operationId: getCounters
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CounterResponse"
    post:
      tags:
      - Counters
      operationId: createCounter
      parameters:
      - name: Idempotency-Key
        in: header
        description: Optional. A new unique value (e.g. a UUID) per action; resend
          the same value only when retrying it.
        schema:
          type: string
          maxLength: 64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCounterRequest"
        required: true
      responses:
        "201":
          description: Counter created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CounterResponse"
        "422":
          description: Idempotency-Key reused
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "409":
          description: Idempotency-Key is already being processed
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Counter creation failed
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ValidationProblem"
  /api/counters/{id}/reset:
    post:
      tags:
      - Counters
      operationId: resetCounter
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "404":
          description: Counter not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "200":
          description: The counter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CounterResponse"
  /api/counters/{id}/increment:
    post:
      tags:
      - Counters
      operationId: incrementCounter
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: Idempotency-Key
        in: header
        description: Optional. A new unique value (e.g. a UUID) per action; resend
          the same value only when retrying it.
        schema:
          type: string
          maxLength: 64
      responses:
        "404":
          description: Counter not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "422":
          description: Unprocessable Entity
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              examples:
                IdempotencyKeyReused:
                  summary: Idempotency-Key reused
                  description: "The request parameters changed, but the Idempotency-Key\
                    \ was reused."
                  value:
                    title: Unprocessable Entity
                    status: 422
                    detail: Idempotency key f761e38d-c782-4fa1-9252-78d910c50d3a was
                      already used for a different request
                CounterOverflow:
                  summary: Increment would overflow the count
                  description: New count would exceed max of 2147483647
                  value:
                    title: Unprocessable Entity
                    status: 422
                    detail: New count would exceed max of 2147483647
        "409":
          description: Idempotency-Key is already being processed
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "200":
          description: The counter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CounterResponse"
  /api/counters/{id}:
    get:
      tags:
      - Counters
      operationId: getCounter
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "404":
          description: Counter not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "200":
          description: The counter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CounterResponse"
    delete:
      tags:
      - Counters
      operationId: deleteCounter
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "404":
          description: Counter not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "204":
          description: Counter deleted
components:
  schemas:
    CreateCounterRequest:
      type: object
      properties:
        label:
          type: string
          maxLength: 100
          minLength: 1
        step:
          type: integer
          format: int32
          minimum: 1
        start:
          type: integer
          format: int32
      required:
      - label
      - step
    Problem:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
      required:
      - detail
      - instance
      - status
      - title
    ValidationProblem:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
        errors:
          type: object
          additionalProperties:
            type: string
      required:
      - detail
      - errors
      - instance
      - status
      - title
    CounterResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        label:
          type: string
        step:
          type: integer
          format: int32
        start:
          type: integer
          format: int32
        count:
          type: integer
          format: int32
      required:
      - count
      - id
      - label
      - start
      - step